initrd installer: Provide real path to hook function instead of symlink
In the 0.142+deb12u1 version of initramfs-tools there were some changes
made to the copy_file hook function (05e5fb9 and f52ae2d). This caused
the binaries in the initrd installer to not be updated, which led to
some commands failing (e.g. cpio). We must make sure to always provide
a real path to copy_exec/copy_file, NOT a symlink, so the binaries in
the initrd installer are updated properly.